Make workload a tuple of ints instead of floats
If we use floats, the problem is no integer programming.

+++ edarop/model.py
@@ -139,7 +139,7 @@ class InstanceClass:
 
 @dataclass(frozen=True)
 class Workload:
-    values: Tuple[float, ...]
+    values: Tuple[int, ...]
     time_unit: TimeUnit
